refactor(LottoMachine): lottoCalculator 객체로 변경
